+++ ArWikiCats/main_processers/main_resolve.py
@@ -7,9 +7,9 @@ from __future__ import annotations
 import functools
 from dataclasses import dataclass
 
-from . import event2bot, event_lab_bot
+from . import event2bot, event_lab_bot, nat_men_pattern
 from .labs_years import LabsYears
-from . import country_time_pattern, nat_pattern
+from . import country_time_pattern, nat_women_time_pattern
 from ..config import app_settings
 from ..fix import fixtitle
 from ..make_bots.co_bots import filter_en
@@ -52,11 +52,11 @@ def resolve_label(category: str) -> CategoryResult:
     start_ylab = ""
     from_match = False
     if not category_lab:
-        category_lab = country_time_pattern.get_label(changed_cat)
+        category_lab = country_time_pattern.get_label(changed_cat) or nat_women_time_pattern.get_label(changed_cat)
         from_match = category_lab != ""
 
     if not category_lab:
-        category_lab = nat_pattern.get_label(changed_cat)
+        category_lab = nat_men_pattern.get_label(changed_cat)
         from_match = category_lab != ""
 
     if not category_lab: