+++ ArWikiCats/translations_formats/__init__.py
@@ -1,15 +1,19 @@
 
-from .DataModel import YearFormatData, FormatData, NormalizeResult, MultiDataFormatterBase, FormatDataDouble
+from .DataModel import YearFormatData, FormatData, NormalizeResult, MultiDataFormatterBase, MultiDataFormatterBaseV2, FormatDataDouble, FormatDataV2, MultiDataFormatterDataDouble, MultiDataFormatterBaseYear
 from .data_with_time import format_year_country_data
 from .data_new_model import format_films_country_data
 from .multi_data import format_multi_data
 
 __all__ = [
+    "MultiDataFormatterBaseYear",
+    "MultiDataFormatterDataDouble",
+    "FormatDataV2",
     "FormatDataDouble",
     "YearFormatData",
     "FormatData",
     "NormalizeResult",
     "MultiDataFormatterBase",
+    "MultiDataFormatterBaseV2",
     "format_year_country_data",
     "format_films_country_data",
     "format_multi_data",